# ConvergenceService/wwwroot/swagger-original.json
{
    "swagger": "2.0",
    "info": {
        "description": "",
        "version": "",
        "title": "TradingView REST API Specification for Brokers"
    },
    "host": "rest-demo.tradingview.com",
    "basePath": "/tradingview/v1/",
...
    "schemes": [
        "https"
    ],
    "consumes": [
        "application/x-www-form-urlencoded"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/symbol_info": {
            "get": {
                "tags": [
                    "Trading Panel bridge (broker data & orders)",
                    "Trading Terminal integration"
                ],
                "description": "Get a list of all instruments",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "List of instruments",
                        "schema": {
                            "$ref": "#/definitions/SymbolInfoArrays"
                        }
                    }
                },
                "security": [
                    {
                        "oauth": []
                    }
                ]
            }
        },
        "/history": {
            "get": {
                "tags": [
                    "Trading Panel bridge (broker data & orders)",
                    "Trading Terminal integration"
                ],
                "description":
                    "Bars request. You can find examples in the [documentation](https://github.com/tradingview/charting_library/wiki/UDF#bars).",
                "parameters": [
                    {
                        "name": "symbol",
                        "in": "query",
                        "description": "Symbol name or ticker",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "resolution",
                        "in": "query",
                        "description":
                            "Symbol resolution. Possible resolutions are daily (`1D`, `2D` ... ), weekly (`1W`, `2W` ...), monthly (`1M`, `2M`...) and an intra-day resolution &ndash; minutes(`1`, `2` ...).",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Unix timestamp (UTC) of the leftmost required bar, including `from`.",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "Unix timestamp (UTC) of the rightmost required bar, including `to`.",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "countback",
                        "in": "query",
                        "description":
                            "Number of bars (higher priority than `from`) starting with `to`. If `countback` is set, `from` should be ignorred. It is used only by tradingview.com, Trading Terminal will never use it.",
                        "required": false,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description":
                            "Response is expected to be an object with properties listed below. Each property is treated as a table column",
                        "schema": {
                            "$ref": "#/definitions/BarsArrays"
                        }
                    }
                },
                "security": [
                    {
                        "oauth": []
                    }
                ]
            }
        },
...
        "/streaming": {
            "get": {
                "tags": [
                    "Trading Panel bridge (broker data & orders)"
                ],
                "description":
                    "Stream of prices. Server constantly keeps the connection alive. If the connection is broken the server constantly tries to restore it. Transfer mode is 'chunked encoding'. The data feed should set 'Transfer-Encoding: chunked' and make sure that all intermediate proxies are set to use this mode. All messages are finished with '\\n'. Streaming data should contain real-time only. It shouldn't contain snapshots of data.",
                "parameters": [],
                "responses": {
                    "200": {
                        "description":
                            "Data feed should provide ticks (trades, asks, bids) and daily bars.\n\nIf there is no trades the data feed should set trades to bids.\n\nIf there is only ask/bid implementation you must also set the trade (same as bid but it's size must be >= 1).\n\nSize for trades is always >= 1 except for a correction. In case of correction size can be 0.\n\nAll times should be UNIX time UTC.\n\nDaily bars are required if they cannot be built from ticks (has_dwm should be set to true in the symbol information).\n\nFields for asks, bids and trades: `id`, `p`, `s` (optional for asks and bids), `t`, `f`.\n\nFields for daily bars: `id`, `t`, `o`, `h`, `l`, `c`, `v`.\n\nMessages:\n1. trade `{\"id\":\"symbol\",\"p\":price,\"s\":size,\"t\":time}`\n2. ask `{\"id\":\"symbol\",\"p\":price,\"s\":size,\"t\":time,\"f\":\"a\"}`\n3. bid `{\"id\":\"symbol\",\"p\":price,\"s\":size,\"t\":time,\"f\":\"b\"}`\n4. daily bar `{\"id\":\"symbol\",\"o\":open,\"h\":high,\"l\":low,\"c\":close,\"v\":volume,\"t\":time,\"f\":\"d\"}`\n",
                        "schema": {
                            "$ref": "#/definitions/inline_response_200_14"
                        }
                    }
                },
                "security": [
                    {
                        "oauth": []
                    }
                ]
            }
        }
    },
    "securityDefinitions": {
        "oauth": {
            "description":
                "Look for more request details in /authorize request. TradingView will set Authorization header to 'Bearer ' + access_token for all requests with authorization.",
            "type": "oauth2",
            "tokenUrl": "http://staging-paper-rest-api.tradingview.com/tradingview/v1/authorize",
            "flow": "password",
            "scopes": {
                "general": "permission to perform all requests"
            }
        }
    },
    "definitions": {
...
        "SymbolInfoArrays": {
            "type": "object",
            "required": [
                "description",
                "exchange-listed",
                "exchange-traded",
                "minmovement",
                "pricescale",
                "session-regular",
                "symbol",
                "timezone"
            ],
            "properties": {
                "symbol": {
                    "type": "array",
                    "description":
                        "It's name of a symbol. It is a string which your users will see. Also, it will be used for data requests if you are not using tickers.",
                    "items": {
                        "type": "string"
                    }
                },
                "description": {
                    "type": "array",
                    "description": "Description of a symbol. Will be printed in chart legend for this symbol.",
                    "items": {
                        "type": "string"
                    }
                },
                "exchange-listed": {
                    "type": "array",
                    "description": "Short name of exchange where this symbol is listed",
                    "items": {
                        "type": "string"
                    }
                },
                "exchange-traded": {
                    "type": "array",
                    "description": "Short name of exchange where this symbol is traded",
                    "items": {
                        "type": "string"
                    }
                },
                "minmovement": {
                    "type": "array",
                    "description": "Minimal integer price change",
                    "items": {
                        "type": "number"
                    }
                },
                "minmov2": {
                    "type": "array",
                    "description": "It's a number for complex price formatting cases",
                    "items": {
                        "type": "number"
                    }
                },
                "fractional": {
                    "type": "array",
                    "description":
                        "Boolean showing whether this symbol wants to have complex price formatting (see minmov2) or not",
                    "items": {
                        "type": "boolean"
                    }
                },
                "pricescale": {
                    "type": "array",
                    "description": "MinimalPossiblePriceChange = minmovement / pricescale",
                    "items": {
                        "type": "number"
                    }
                },
                "has-intraday": {
                    "type": "array",
                    "description":
                        "Boolean showing whether symbol has intraday (minutes) history data. If it's false then all buttons for intradays resolutions will be disabled when this symbol is active in chart. If it is set to true, all resolutions that are supplied directly by the datafeed must be provided in intraday_multipliers array.",
                    "items": {
                        "type": "boolean"
                    }
                },
                "has-no-volume": {
                    "type": "array",
                    "description": "Boolean showing whether symbol has volume data or not",
                    "items": {
                        "type": "boolean"
                    }
                },
                "type": {
                    "type": "array",
                    "description": "Symbol type (forex/stock etc.)",
                    "items": {
                        "type": "string"
                    }
                },
                "ticker": {
                    "type": "array",
                    "description":
                        "It's an unique identifier for this symbol in your symbology. If you specify this property then its value will be used for all data requests for this symbol. ticker is treated to be equal to symbol if not specified explicitly.",
                    "items": {
                        "type": "string"
                    }
                },
                "timezone": {
                    "type": "array",
                    "description":
                        "Exchange timezone for this symbol. We expect to get name of time zone in olsondb format",
                    "items": {
                        "type": "string"
                    }
                },
                "session-regular": {
                    "type": "array",
                    "description":
                        "Trading hours for this symbol. See the [Trading Sessions](https://github.com/tradingview/charting_library/wiki/Trading-Sessions) article to know more details.",
                    "items": {
                        "type": "string"
                    }
                },
                "supported-resolutions": {
                    "type": "array",
                    "description":
                        "An array of resolutions which should be enabled in resolutions picker for this symbol. Each item of an array is expected to be a string.",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "has-daily": {
                    "type": "array",
                    "description":
                        "The boolean value showing whether datafeed has its own D resolution bars or not. If has_daily = false then Charting Library will build respective resolutions from intraday by itself. If not, then it will request those bars from datafeed.",
                    "items": {
                        "type": "boolean"
                    }
                },
                "intraday-multipliers": {
                    "type": "array",
                    "description":
                        "It is an array containing intraday resolutions (in minutes) the datafeed wants to build by itself. E.g., if the datafeed reported he supports resolutions [\"1\", \"5\", \"15\"], but in fact it has only 1 minute bars for symbol X, it should set intraday_multipliers of X = [1]. This will make Charting Library to build 5 and 15 resolutions by itself.",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "has-weekly-and-monthly": {
                    "type": "array",
                    "description":
                        "The boolean value showing whether datafeed has its own W and M resolution bars or not. If has_weekly_and_monthly = false then Charting Library will build respective resolutions from D by itself. If not, then it will request those bars from datafeed.",
                    "items": {
                        "type": "boolean"
                    }
                }
            },
            "description":
                "SymbolInfo is an object containing symbols metadata. Each value of this object is an array of values which size is equal to symbols count or a single value that is applied to all symbols.",
            "example": {
                "symbol": [
                    "symbol",
                    "symbol"
                ],
                "supported-resolutions": [
                    [
                        "supported-resolutions",
                        "supported-resolutions"
                    ],
                    [
                        "supported-resolutions",
                        "supported-resolutions"
                    ]
                ],
                "ticker": [
                    "ticker",
                    "ticker"
                ],
                "has-daily": [
                    true,
                    true
                ],
                "minmov2": [
                    6.02745618307040320615897144307382404804229736328125,
                    6.02745618307040320615897144307382404804229736328125
                ],
                "has-weekly-and-monthly": [
                    true,
                    true
                ],
                "timezone": [
                    "timezone",
                    "timezone"
                ],
                "fractional": [
                    true,
                    true
                ],
                "description": [
                    "description",
                    "description"
                ],
                "intraday-multipliers": [
                    [
                        "intraday-multipliers",
                        "intraday-multipliers"
                    ],
                    [
                        "intraday-multipliers",
                        "intraday-multipliers"
                    ]
                ],
                "type": [
                    "type",
                    "type"
                ],
                "has-no-volume": [
                    true,
                    true
                ],
                "exchange-listed": [
                    "exchange-listed",
                    "exchange-listed"
                ],
                "has-intraday": [
                    true,
                    true
                ],
                "exchange-traded": [
                    "exchange-traded",
                    "exchange-traded"
                ],
                "minmovement": [
                    0.80082819046101150206595775671303272247314453125,
                    0.80082819046101150206595775671303272247314453125
                ],
                "pricescale": [
                    1.46581298050294517310021547018550336360931396484375,
                    1.46581298050294517310021547018550336360931396484375
                ],
                "session-regular": [
                    "session-regular",
                    "session-regular"
                ]
            }
        },
        "BarsArrays": {
            "type": "object",
            "required": [
                "s"
            ],
            "properties": {
                "s": {
                    "type": "string",
                    "description": "Status code.",
                    "enum": [
                        "ok",
                        "error",
                        "no_data"
                    ]
                },
                "errmsg": {
                    "type": "string",
                    "description": "Error message. Should be provided if s = \"error\""
                },
                "nb": {
                    "type": "number",
                    "description":
                        "unix time of the next bar if there is no data (status code is no_data) in the requested period (optional)"
                },
                "t": {
                    "type": "array",
                    "description":
                        "bar time, unix timestamp (UTC). Daily bars should only have the date part, time should be 0.",
                    "items": {
                        "type": "number"
                    }
                },
                "o": {
                    "type": "array",
                    "description": "open price",
                    "items": {
                        "type": "number"
                    }
                },
                "h": {
                    "type": "array",
                    "description": "high price",
                    "items": {
                        "type": "number"
                    }
                },
                "l": {
                    "type": "array",
                    "description": "low price",
                    "items": {
                        "type": "number"
                    }
                },
                "c": {
                    "type": "array",
                    "description": "close price",
                    "items": {
                        "type": "number"
                    }
                },
                "v": {
                    "type": "array",
                    "description": "volume",
                    "items": {
                        "type": "number"
                    }
                }
            },
            "description": "Bars data.",
            "example": {
                "s": "ok",
                "c": [
                    2.3021358869347654518833223846741020679473876953125,
                    2.3021358869347654518833223846741020679473876953125
                ],
                "nb": 0.80082819046101150206595775671303272247314453125,
                "t": [
                    6.02745618307040320615897144307382404804229736328125,
                    6.02745618307040320615897144307382404804229736328125
                ],
                "v": [
                    7.061401241503109105224211816675961017608642578125,
                    7.061401241503109105224211816675961017608642578125
                ],
                "h": [
                    5.962133916683182377482808078639209270477294921875,
                    5.962133916683182377482808078639209270477294921875
                ],
                "errmsg": "errmsg",
                "l": [
                    5.63737665663332876420099637471139430999755859375,
                    5.63737665663332876420099637471139430999755859375
                ],
                "o": [
                    1.46581298050294517310021547018550336360931396484375,
                    1.46581298050294517310021547018550336360931396484375
                ]
            }
        },
...
        "inline_response_200_14": {
            "required": [
                "id",
                "t"
            ],
            "properties": {
                "id": {
                    "type": "string",
                    "description": "Symbol"
                },
                "p": {
                    "type": "number",
                    "description": "Price"
                },
                "s": {
                    "type": "number",
                    "description": "Size"
                },
                "t": {
                    "type": "number",
                    "description": "Time"
                },
                "f": {
                    "type": "string",
                    "description": "Type (a - ask, b - bid, t - trade, d - daily bar). Default value is `t` (trade).",
                    "enum": [
                        "a",
                        "b",
                        "d",
                        "t"
                    ]
                },
                "o": {
                    "type": "number",
                    "description": "Open"
                },
                "h": {
                    "type": "number",
                    "description": "High"
                },
                "l": {
                    "type": "number",
                    "description": "Low"
                },
                "c": {
                    "type": "number",
                    "description": "Close"
                },
                "v": {
                    "type": "number",
                    "description": "Volume"
                }
            },
            "example": {
                "p": 0.80082819046101150206595775671303272247314453125,
                "s": 6.02745618307040320615897144307382404804229736328125,
                "c": 7.061401241503109105224211816675961017608642578125,
                "t": 1.46581298050294517310021547018550336360931396484375,
                "f": "a",
                "v": 9.301444243932575517419536481611430644989013671875,
                "h": 5.63737665663332876420099637471139430999755859375,
                "id": "id",
                "l": 2.3021358869347654518833223846741020679473876953125,
                "o": 5.962133916683182377482808078639209270477294921875
            }
        },
...
    },
...
}